dejé todos los archivos que tenían subclases, en sus propios archivo dentro de una carpeta con el nombre de la clase
