+++ Tarea/clasebebida/CocaCola.java
@@ -1,0 +1,11 @@
+package Tarea.clasebebida;
+
+class CocaCola extends Bebida {
+    public CocaCola(int ac) {
+        super(ac);
+    }
+
+    public String beber() {
+        return new String("cocacola");
+    }
+}
