+++ Tarea/clasebebida/Pepsi.java
@@ -1,0 +1,11 @@
+package Tarea.clasebebida;
+
+class Pepsi extends Bebida {
+    public Pepsi(int ap) {
+        super(ap);
+    }
+
+    public String beber() {
+        return new String("pepsi");
+    }
+}
