+++ Tarea/clasebebida/Sprite.java
@@ -1,0 +1,11 @@
+package Tarea.clasebebida;
+
+class Sprite extends Bebida {
+    public Sprite(int ab) {
+        super(ab);
+    }
+
+    public String beber() {
+        return new String("sprite");
+    }
+}
