+++ Tarea/clasedulces/Calugas.java
@@ -1,0 +1,11 @@
+package Tarea.clasedulces;
+
+class Calugas extends Dulces{
+    public Calugas(int ac) {
+        super(ac);
+    }
+
+    public String dulces() {
+        return new String("calugas");
+    }
+}
