+++ Tarea/clasedulces/Oreos.java
@@ -1,0 +1,12 @@
+package Tarea.clasedulces;
+
+
+class Oreos extends Dulces{
+    public Oreos(int ac) {
+        super(ac);
+    }
+
+    public String dulces() {
+        return new String("oreos");
+    }
+}
