+++ Tarea/clasedulces/Serranita.java
@@ -1,0 +1,11 @@
+package Tarea.clasedulces;
+
+class Serranita extends Dulces{
+    public Serranita(int ac) {
+        super(ac);
+    }
+
+    public String dulces() {
+        return new String("serranita");
+    }
+}
