+++ Tarea/clasemoneda/Moneda.java
@@ -1,0 +1,12 @@
+package Tarea.clasemoneda;
+
+public abstract class Moneda {
+    public Moneda getSerie() {
+        return this;
+    }
+
+    public abstract int getValor();
+
+    public Moneda() {
+    }
+}
