+++ Tarea/clasemoneda/Moneda100.java
@@ -1,0 +1,11 @@
+package Tarea.clasemoneda;
+
+public class Moneda100 extends Moneda {
+    public int getValor() {
+        return 100;
+    }
+
+    public Moneda100() {
+        super();
+    }
+}
