+++ Tarea/clasemoneda/Moneda1000.java
@@ -1,0 +1,11 @@
+package Tarea.clasemoneda;
+
+public class Moneda1000 extends Moneda {
+    public int getValor() {
+        return 1000;
+    }
+
+    public Moneda1000() {
+        super();
+    }
+}
