+++ Tarea/clasemoneda/Moneda1500.java
@@ -1,0 +1,11 @@
+package Tarea.clasemoneda;
+
+public class Moneda1500 extends Moneda {
+    public int getValor() {
+        return 1500;
+    }
+
+    public Moneda1500() {
+        super();
+    }
+}
