+++ Tarea/clasemoneda/Moneda500.java
@@ -1,0 +1,11 @@
+package Tarea.clasemoneda;
+
+public class Moneda500 extends Moneda {
+    public int getValor() {
+        return 500;
+    }
+
+    public Moneda500() {
+        super();
+    }
+}
